+++ mayan/apps/navigation/templatetags/navigation_tags.py
@@ -108,6 +108,11 @@ def navigation_resolve_menus(context, names, source=None, sort_results=None):
     return result
 
 
+@register.simple_tag()
+def navigation_source_column_get_absolute_url(source_column, obj):
+    return source_column.get_absolute_url(obj=obj)
+
+
 @register.simple_tag(takes_context=True)
 def resolve_link(context, link):
     # This can be used to resolve links or menus too
@@ -121,4 +126,3 @@ def navigation_source_column_resolve(context, column):
         return result
     else:
         return ''
-